Remove HybridEngineWriteBatch
