+++ components/backup-stream/src/observer.rs
@@ -296,7 +296,7 @@ mod tests {
         // Test region out of range won't be added to observe list.
         let r = fake_region(43, b"0010", b"0042");
         let mut ctx = ObserverContext::new(&r);
-        o.on_role_change(&mut ctx, &RoleChange::new(StateRole::Leader));
+        o.on_role_change(&mut ctx, &RoleChange::new_for_test(StateRole::Leader));
         let task = rx.recv_timeout(Duration::from_millis(20));
         assert!(task.is_err(), "it is {:?}", task);
         assert!(!subs.is_observing(43));
@@ -311,7 +311,7 @@ mod tests {
         // Test give up subscripting when become follower.
         let r = fake_region(42, b"0008", b"0009");
         let mut ctx = ObserverContext::new(&r);
-        o.on_role_change(&mut ctx, &RoleChange::new(StateRole::Follower));
+        o.on_role_change(&mut ctx, &RoleChange::new_for_test(StateRole::Follower));
         let task = rx.recv_timeout(Duration::from_millis(20));
         assert_matches!(
             task,
@@ -333,7 +333,7 @@ mod tests {
             RegionChangeEvent::Update(RegionChangeReason::Split),
             StateRole::Leader,
         );
-        o.on_role_change(&mut ctx, &RoleChange::new(StateRole::Leader));
+        o.on_role_change(&mut ctx, &RoleChange::new_for_test(StateRole::Leader));
         let task = rx.recv_timeout(Duration::from_millis(20));
         assert!(task.is_err(), "it is {:?}", task);
     }
