+++ components/engine_traits/src/write_batch.rs
@@ -1,7 +1,5 @@
 // Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
 
-use kvproto::metapb;
-
 use crate::{errors::Result, options::WriteOptions};
 
 /// Engines that can create write batches
@@ -138,8 +136,4 @@ pub trait WriteBatch: Mutable {
     fn merge(&mut self, src: Self) -> Result<()>
     where
         Self: Sized;
-
-    /// It declares that the following consecutive write will be within this
-    /// region.
-    fn prepare_for_region(&mut self, _: &metapb::Region) {}
 }
