+++ components/raftstore/src/coprocessor/read_write/write_batch.rs
@@ -13,11 +13,9 @@ pub trait WriteBatchObserver: Send {
 /// additional methods to specify which region the write operations belong to.
 // TODO: May be we can unified it with `CmdObserver`?
 pub trait ObservableWriteBatch: WriteBatch + Send {
-    // It declares that the following consecutive write will be within this
-    // region.
-    // TODO: move `prepare_for_region` from `WriteBatch` to here.
-    // fn prepare_for_region(&mut self, _: CacheRegion);
-
+    /// It declares that the following consecutive write will be within this
+    /// region.
+    fn prepare_for_region(&mut self, region: &metapb::Region);
     /// Commit the WriteBatch with the given options and sequence number.
     fn write_opt_seq(&mut self, opts: &WriteOptions, seq_num: u64);
     /// It is called after a write operation is finished.
@@ -37,6 +35,12 @@ impl<WB> WriteBatchWrapper<WB> {
         Self {
             write_batch,
             observable_write_batch,
+        }
+    }
+
+    pub fn prepare_for_region(&mut self, region: &metapb::Region) {
+        if let Some(w) = self.observable_write_batch.as_mut() {
+            w.prepare_for_region(region)
         }
     }
 }
@@ -112,12 +116,6 @@ impl<WB: WriteBatch> WriteBatch for WriteBatchWrapper<WB> {
 
     fn merge(&mut self, _: Self) -> Result<()> {
         unimplemented!("WriteBatchWrapper does not support merge")
-    }
-
-    fn prepare_for_region(&mut self, region: &metapb::Region) {
-        if let Some(w) = self.observable_write_batch.as_mut() {
-            w.prepare_for_region(region)
-        }
     }
 }
